Implement equals button, adjust updateDisplay
Function updateDisplay changed to take in an argument, newDisplay.
Display will be updated with newDisplay. Display was previously updated
with value of currentOperand but this was too restrictive for how the
logic is supposed to work in certain cases.

Global boolean operatorSelected added to better handle what should
happen to the display when a number is pressed (clear vs append).

Function operate adjusted to convert strings to numbers before
arithmetic and back to strings when returning result.